REFACTOR: Moved certificate related functions to certificate folder
